feat: extract shared utilities package
