+++ packages/platform-core/src/createShop.ts
@@ -10,7 +10,8 @@ import { ulid } from "ulid";
 import { z } from "zod";
 import { validateShopName } from "./shops";
 import { copyTemplate, loadBaseTokens } from "./createShop/utils";
-import { slugify, genSecret, fillLocales } from "./utils";
+import { slugify, genSecret } from "@shared-utils";
+import { fillLocales } from "./utils";
 import { loadThemeTokensNode } from "./themeTokens";
 import { nowIso } from "@shared/date";
 import { defaultFilterMappings } from "./defaultFilterMappings";
