+++ packages/platform-core/src/utils/index.ts
@@ -1,3 +1,3 @@
 export { getShopFromPath } from "./getShopFromPath";
 export { replaceShopInPath } from "./replaceShopInPath";
-export { slugify, genSecret, fillLocales } from "./string";
+export { fillLocales } from "./string";
