+++ packages/platform-core/src/utils/string.ts
@@ -1,21 +1,6 @@
 // packages/platform-core/src/utils/string.ts
 import { LOCALES } from "@i18n/locales";
 import type { Locale } from "@types";
-import { randomBytes } from "crypto";
-
-/** Convert a string into a URL-friendly slug. */
-export function slugify(str: string): string {
-  return str
-    .trim()
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/^-+|-+$/g, "");
-}
-
-/** Generate a random secret represented as a hexadecimal string. */
-export function genSecret(bytes = 16): string {
-  return randomBytes(bytes).toString("hex");
-}
 
 /**
  * Ensure all locales have a value, filling in missing entries with a fallback.
