+++ packages/shared-utils/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@acme/shared-utils",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "test": "jest --ci --runInBand --detectOpenHandles --config ../../jest.config.cjs"
+  }
+}
