+++ packages/shared-utils/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./fetchJson";
+export * from "./string";
