+++ packages/ui/src/utils/index.ts
@@ -1,6 +1,5 @@
 export * from "./boxProps";
 export * from "./cn";
 export * from "./drawerWidth";
-export * from "./fetchJson";
 export * from "./multilingual";
 export * from "./toggleItem";
